Image Upload and Invoice Generation
New parser that can handle file requests (multer)
Added option to get invoice with authorization
PDF Kit to create pdf

+++ controllers/admin.js
@@ -1,5 +1,6 @@
 const Product = require('../models/product');
 const { validationResult } = require('express-validator');
+const fileHelper = require('../util/file');
 
 exports.getProducts = (req, res,next) => {
     Product.find({ userId: req.user._id })
@@ -33,16 +34,34 @@ exports.getAddProduct = (req, res, next) => {
 
 // Saves the new product with post request
 exports.postAddProduct = (req, res, next) => {
-    const { title, imageUrl, price, description } = req.body
+    const { title, price, description } = req.body;
+    
+    const image = req.file;
+
+    if(!image) {
+        return res.status(422).render('admin/edit-product', {
+            pageTitle: 'Add Product',
+            path: '/admin/add-product',
+            product: {
+                title: title,
+                price: price,
+                description: description
+            },
+            editing: false,
+            hasError: true,
+            errorMessage: "Invalid file uploaded",
+            validationErrors: []
+        })
+    }
 
     const errors = validationResult(req)
+
     if(!errors.isEmpty()) {
         return res.status(422).render('admin/edit-product', {
             pageTitle: 'Add Product',
             path: '/admin/add-product',
             product: {
                 title: title,
-                imageUrl: imageUrl,
                 price: price,
                 description: description
             },
@@ -52,6 +71,8 @@ exports.postAddProduct = (req, res, next) => {
             validationErrors: errors.array()
         })
     }
+
+    const imageUrl = `\\${image.path}`
 
     const product = new Product({
         title: title,
@@ -64,7 +85,6 @@ exports.postAddProduct = (req, res, next) => {
     product.save()
         .then(result => {
             console.log('New Product inserted...')
-            console.log(result)
             res.redirect('/admin/products')
         })
         .catch(err => {
@@ -76,7 +96,9 @@ exports.postAddProduct = (req, res, next) => {
 
 // Return the page to edit existing product
 exports.getEditProduct = (req, res, next) => {
+    
     const productId = req.params.productId
+    
     Product.findOne({ _id: productId, userId: req.user._id })
         .then(product => {
             if(!product) {
@@ -101,7 +123,10 @@ exports.getEditProduct = (req, res, next) => {
 
 // Saves the edited product with post request
 exports.postEditProduct = (req, res, next) => {
-    const { id, title, imageUrl, price, description } = req.body
+
+    const { id, title, price, description } = req.body
+
+    const image = req.file;
 
     const errors = validationResult(req)
     if(!errors.isEmpty()) {
@@ -111,7 +136,6 @@ exports.postEditProduct = (req, res, next) => {
             product: {
                 _id: id,
                 title: title,
-                imageUrl: imageUrl,
                 price: price,
                 description: description
             },
@@ -128,9 +152,12 @@ exports.postEditProduct = (req, res, next) => {
                 return res.redirect('/')
             }
             product.title = title;
-            product.imageUrl = imageUrl;
             product.price = price;
             product.description = description;
+            if(image) {
+                fileHelper.deleteFile(product.imageUrl);
+                product.imageUrl = `\\${image.path}`;
+            }
             return product.save()
         })
         .then(result => {
@@ -147,18 +174,20 @@ exports.postEditProduct = (req, res, next) => {
 // Delete the product with id
 exports.postDeleteProduct = (req, res, next) => {
     const id = req.body.productId;
-    Product.deleteOne({ _id: id, userId: req.user._id })
-        .then(result => {
-            console.log(result)
-            if(result.deletedCount > 0) {
-                console.log("Product deleted...")
-                return res.redirect('/admin/products')
-            }
-            res.redirect('/')
-        })
-        .catch(err => {
-            const error = new Error(err);
-            error.httpStatusCode = 500;
-            return next(error);
-        })
+    Product.findById(id).then(product => {
+        if(!product) {
+            return next(new Error('Product not found.'))
+        }
+        fileHelper.deleteFile(product.imageUrl);
+        return Product.deleteOne({ _id: id, userId: req.user._id });
+    })
+    .then(() => {
+        console.log("Product deleted...")
+        res.redirect('/admin/products')
+    })
+    .catch(err => {
+        const error = new Error(err);
+        error.httpStatusCode = 500;
+        return next(error);
+    })
 }
